feat: App 컴포넌트 최적화
