+++ src/@lib/provider/AuthContextProvider.tsx
@@ -1,0 +1,50 @@
+import { createContext, PropsWithChildren, useContext, useState } from "react";
+import { useNotificationContext } from "./NotificationProvider";
+import { useCallback, useMemo } from "../hooks";
+
+interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+interface Auth {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+const AuthContext = createContext<Auth | undefined>(undefined);
+
+export const useAuthContext = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
+
+export default function AuthContextProvider({ children }: PropsWithChildren) {
+  const [user, setUser] = useState<User | null>(null);
+
+  const { addNotification } = useNotificationContext();
+
+  const login = useCallback((email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    addNotification("성공적으로 로그인되었습니다", "success");
+  }, []);
+
+  const logout = useCallback(() => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  }, []);
+
+  const contextValue = useMemo(
+    () => ({ user, login, logout }),
+    [user, login, logout],
+  );
+
+  return (
+    <AuthContext.Provider value={contextValue}>{children}</AuthContext.Provider>
+  );
+}
