+++ src/@lib/provider/ThemeContextProvider.tsx
@@ -1,0 +1,35 @@
+import { createContext, PropsWithChildren, useContext, useState } from "react";
+import { useCallback, useMemo } from "../hooks";
+
+const ThemeContext = createContext({
+  theme: "light",
+  toggleTheme: () => {},
+});
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
+
+export const ThemeContextProvider = ({ children }: PropsWithChildren) => {
+  const [theme, setTheme] = useState<"light" | "dark">("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  }, []);
+
+  const contextValue = useMemo(() => ({ theme, toggleTheme }), [theme]);
+
+  return (
+    <ThemeContext.Provider value={contextValue}>
+      <div
+        className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+      >
+        {children}
+      </div>
+    </ThemeContext.Provider>
+  );
+};
